fixing int overflow

+++ app/src/main/java/com/example/sipcalculator/model/InputDataModel.kt
@@ -6,10 +6,10 @@ import kotlinx.android.parcel.Parcelize
 @Parcelize
 data class InputDataModel(
     val years: Int,
-    val monthlyAmount: Int,
+    val monthlyAmount: Double,
     val expectedReturns: Double,
-    val lumpsum: Int = 0,
+    val lumpsum: Double = 0.0,
     val inflationRate: Double = 0.0,
-    val topupAmount: Int = 0,
+    val topupAmount: Double = 0.0,
     val topupRate: Double = 0.0
 ): Parcelable
